hou/requirements.txt
Flask>=3.0.0
Flask-RESTful>=0.3.10
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.25
alembic>=1.13.1
Flask-Migrate>=4.0.5
PyMySQL>=1.1.0
bcrypt>=4.1.2
pydantic>=2.5.0
python-docx>=1.1.0
pypdf>=4.0.0
openpyxl>=3.1.2
psutil>=5.9.0
APScheduler>=3.10.4
oss2>=2.18.4
openai==1.12.0
numpy>=1.24.0
python-dotenv>=1.0.0
flasgger>=0.9.7.1
cryptography>=42.0.0
